hsmd: Rename hsmd_ready_channel to hsmd_setup_channel
When BOLT-2 renamed funding_locked to channel_ready the hsmd message
named ready_channel became very confusing.  Change it to setup_channel
instead to not confuse developers.not trigger the wrong association.

Because the hsmd wire messages are parsed by number this change is
forward and backwards compatible.

Changelog-Changed: hsmd: Renamed hsmd_ready_channel to hsmd_setup_channel because of the BOLT-2 rename of funding_locked to channel_ready