channeld: split get_per_commitment_point usesinto separate functions
Changelog-None: channeld internal

This factoring makes it much clearer which callers need the pubkey and
which need the old_secret.

It does not alter the unfortunate semantic that fetching future
pubkeys beyond the next commit is an error (because the signer will
only release secrets for revoked commitments).  Using HSM_VERSION 6
will fix this semantic in hsmd.